+++ web/src/components/GrIntro.vue
