+++ web/src/components/GrLegend.vue
