+++ web/src/components/GrVisLegend.vue
